% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LeesApprox}
\alias{LeesApprox}
\title{Title}
\usage{
LeesApprox(FVec, ngtg, maxsd, binwidth, M, Linf, K, t0, LFS, L5, Vmaxlen,
  LinfCV, maxage)
}
\arguments{
\item{x}{An integer vector}
}
\description{
Description
}
