% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeesApproxTMB.R
\name{LeesApproxTMB}
\alias{LeesApproxTMB}
\title{Lees Approximation interpolation function}
\usage{
LeesApproxTMB(
  FVec,
  ngtg,
  maxsd,
  binwidth,
  M,
  Linf,
  K,
  t0,
  LFS,
  L5,
  Vmaxlen,
  LinfCV,
  maxage
)
}
\description{
Primarily used to verify that output from TMB will match that from Rcpp to numerical precision.
}
